Revert "Do not reorder side effect nodes #110"
