+++ src/constants.ts
@@ -7,9 +7,6 @@ export const jsx: ParserPlugin = 'jsx';
 
 export const newLineCharacters = '\n\n';
 
-export const chunkSideEffectNode = 'side-effect-node';
-export const chunkSideOtherNode = 'other-node';
-
 /*
  * Used to mark the position between RegExps,
  * where the not matched imports should be placed
