+++ src/utils/get-sorted-nodes-group.ts
@@ -1,5 +1,4 @@
 import { Import, ImportDeclaration } from '@babel/types';
-
 import { naturalSort } from '../natural-sort';
 import { PrettierOptions } from '../types';
