--- tests/ImportsSeparated/import-export-in-between-side-effect.ts
@@ -1,21 +1,0 @@
-import "side-effect";
-import threeLevelRelativePath from "../../../threeLevelRelativePath";
-import sameLevelRelativePath from "./sameLevelRelativePath";
-import thirdParty from "third-party";
-export { random } from './random';
-import c from 'c';
-import oneLevelRelativePath from "../oneLevelRelativePath";
-import otherthing from "@core/otherthing";
-import a from 'a';
-import twoLevelRelativePath from "../../twoLevelRelativePath";
-import component from "@ui/hello";
-export default {
-    title: 'hello',
-};
-import fourLevelRelativePath from "../../../../fourLevelRelativePath";
-import something from "@server/something";
-import x from 'x';
-
-function add(a:number,b:number) {
-  return a + b;
-}
